pyproject.toml
[tool.poetry]
name = "pentest-demo"
version = "0.1.0"
description = ""
authors = ["EINDEX <eindex.me@outlook.com>"]

[[tool.poetry.source]]
name = "pipy"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
jinja2 = "^3.1.2"
pyyaml = "^6.0"
pyjwt = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
